requirements.txt
altgraph==0.17.4
packaging==24.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
speedtest-cli==2.1.3
